create separate logging thread
This extends on the internal log entriy handling to allow for a separate thread to
run through the list of entries for logging.  This is primarily because of the effect
some have reported of logging to certain filesystems whch cause a stall in the writing
routines for logging.

As the actual writing was done within the context of the calling API, then certain
areas could suffer timing constraints. Often showing up with worker stuck messages.
Mostly likely annoying noise in logs but could extend to delayed throughput of data
in extreme cases.

The changes always force adding the line to the entry list and uses a worker wakeup
type method to kick off the writing in the backend thread.  The layout also tries to
be fairly generic in the way mutexes are used within the library.