ipv6 update, mainly for windows build
the windows build had an issue previously with ipv6 so was disable.

This allows for the checking of the right types for win32/64, calling the ipv6 routines
and avoiding the dualstack cases.  The original net API did not really allow for
multiple ports being defined for a single defined listening port (such as ipv4 and ipv6).
This is now extended to allow that.

As windows likes to do things that little bit different, various type checks are added
to improve the straight forward build case.