logging update
Log messages from any root startup were not logged as the log file should be as the changed owner.
This utilises the queue to trap for entries then when ready the committer thread flushes them out.

This meant that the file opening done by calling routines was modified so that the committer thread
dealt with it instead.