avoid the log cycle if no archive is set, unless no file is present.
Some prefer to move/delete log files then signal but don't require the archiving aspect. The
log subsystem did allow for that before the async work but had become broken.  This should now
avoid the re-opening unless the file is no longer present, in which case it assumes a cycle is
needed.

archive set as 0 means cycle to .old
archive unset (defaults to -1) means to reopen if file not present (eg renamed/deleted)
archive set as 1 means cycle with a timestamp.