open up the active client worker cap.  For larger scale client setups.
could do with some reordering of clients to really be useful but for now have quite a
large cap for active clients on workers, and ignore it on worker closing. wakeups are
not restricted either.

I've added the time range warning to the mid cycle worker processing to flag up any
oddities in the log.