make regular reported global stats a bit more flexible
certain stats like global listener counts do not need to be immediately reported on
the log (debug level) or on stats link, so have them update on an upto 10 second
frequency unless they have been changed which reports on the next second.

This allows for a sudden increase or decrease of listeners to not spam the log/link
with a number of updates within a second.