fix a couple of rare fallback cases, one possibly leading to crash
when falling back to file withe private buffers (eg auth intro data, or private queue
buffer copies), the buffer validatation routine was left as-is which would be a source
related routine, so avoid that case.

Also don't terminate a listener if the fallback source does not actually start up, the
lag being 0 and queue size being 0 is ok to continue fallback.