update intro file handling internally
this now uses the sync routine as a separate function and delays the opening/checking of the
file until the first block is read.  This allows us to verify at least the type being a match.
Later the bitrate (possibly other) can be matched as well.

This also allows us to set the intro via /admin/metadata with the intro= or preroll= param.