fix lock imbalance in worker shutdown.
This seems to affect windows more than linux due to lock implementation but needed fixing. Would probably
affect cases of reducing the number of workers and show as a deadlock.

I have resolved the lock imbalance but have also add back in the RW lock for thread cleanup for workers
and log commit thread. It may not be needed but it does lessen the chance of something running into some
cleanup when threads are shutting does in an async way, like the log thread.