rate adjustment allowed for timeout cases
when falling back to file, the average bitrate taken from the source is taken to be the average
incoming rate. In timeout cases, the last N seconds is 0 bytes/s, so allow the rate API to
adjust the range to be less than the whole range. Gives a more accurate figure for fallback
to file