average rate handler update.
make source push a 0 value on each run through to allow for a regular drop of the bitrate if
nothing keeps it up (stalled listeners were not handled previoudly). Ideally we should make it
more time dependant eg every 200ms instead of every read attempt.

change in the way the next block is reused in the circular list. We still have to allow for
inserts as the purge can delete blocks normally.

On a reduce trigger, set a cycle off which initially reduces the sampling to the last N samples
and then keep that number of samples at least for 1 go.  This means that after the reduce, any
residual high samples can be dropped out of the average quickly.