filter out ID3 when streaming for now.
seems like browsers have issues with ID3 embedded in content when not at the beginning. This
poses a problem when a fallback file is used with ID3.

Some slightly processing change in mpeg handling. To prevent a bunch of resync messages
appearing, a failed frame process like for ID3 frames can report the frame size which can
then be skipped or assumed to be 1 byte as before.