slight rework for the global max listeners assessment
we still use the global listeners count, but the stats are updated on the regular update, no inc/dec routines
are used now. There was a bug in the way from fallback file handling was done in this required which could affect
the global max listeners chaeck.  This should be resolved now