make adjustments to timings on stall and queue purge assessment
The listeners need to respond to their positioning for accurate queue sizing, but with large
retries it wa not happening which could lead to a fault if when listeners eventually moved off
the queue.

Here we exptend the assessment time a little, reduce the frequency it is done and reduced some
of the max time frame the listener is tried again.