allow for seconds to be specified in burst, queue and min queue settings
we still allow for the numbers as before but if you want to use seconds then specify like

<burst-size>5s</burst-size>

the awkward bit has been the calculation of bitrate assessment. As there is a variable element
to it (eg relays getting a burst) then the calculation is re-assessed every few seconds.  This
should allow for the default to be specified in secs for any stream, eg 30 secs for queue and
may be 6 seconds for burst and that would apply to 32k or 128k streams.