avoid possible mutex deadlock with run scripts
On fork, a single thread can be running which normally allows for logging but logging requires
taking a lock for a short period of time. If the log lock was taken just before the fork by another
thread then the duplicated process will not have that other thread to unlock.  The af fork handlers
look messy so instead we shuffle around the tests and do no logging in the children.