sync up for remux for mpts
To see if keyframe alignment is better, here we pull packets out of the demux stage and then re-appply
them to produce the stream data to queue.   it turns out that av_read_frame does not react well to
an EAGAIN type situation, a whole punch of packets look to be dropped.  For this I keep a small cache of
incoming pages for feeding into the demux, which pulls out packets for the mux, allowing for the loop
to jump out when the cache gets small enough.

code cleanup should really be done, but wanted to get some further testing done first.