allow for setting a different default charset to utf8.
Originally this was latin1, because it was the best guess for apps at the time. However as utf8
became more univeral then it was used and this was changed.  This can be a problem for some as
they have latin1 contained files and the icecast on there may not be easily configured as some
portals are rather fixed in those regards.

As command line and xml are not a feasible facility for them, then allow the build to revert to
the original charset if they so require.

Now they can run ./configure ICY_CHARSET=latin1
or whatever character set they would set to. Of course the stats engine still uses utf8 and the
conversion is done from their character set to UTF8.