small rework for banned IPs and XFF
the XFF part was done far too late for IP checks. The existing IP acceptance check is still
done there but another is needed after headers come in, as that is where the XFF can be done at
the earliest. The XFF is now brought forward to the header parsing stage and rechecked if
needed.