locking and scheduling update
Just to help in locking of shared settings, some code re-arrangement occurred to avoid any stale
info and maybe take certain locks less in the common cases. I do not think this has really shown
to be a general problem but compilers and platforms do change over time.

There is a scheduling update for retrying listeners, not a huge issue and may be indicating a
wider problem on the system but the tuning was sensible.

Global listeners count was getting stuffed in certain fallback cases. Legacy code that should of
been removed.