fix indefinte auth url loop if auth server keeps redirecting
Default to 1 redirection as you could have some load balancing as a
valid case but if the libcurl default is indefinite so could cause a
stall which leads to other problems. Allow for a single redirection by
default but only allow 0-10. Anything like 10 is silly and is just
wasting critical time.  The option "redirect_limit" can be specified to
change the default, which may include 0 for no redirecton.