fix stale IP data from keepalive, should fix #348
When the client was reset for keepalive and adding back on the worker, any
forwarded IP used in the initial connection stayed around and as it was not
matching the forward for addresses, did not get changed on the subsequent go.

Here we pull out the peer query parts so that it can be called by either
the connection init or reset.