another extention on config locking api. keeps compatability
Uses the refcount on a mount info block to allow for lookup and
holding while dropping the config lock. Allows the lock scopes to
be reduced if a certain amount of processing is to be done.

Added config_lock_mount. Like the config_find_mount, returns a mount
struct but this will increase the refcount on it to avoid deletion if
the config is updated but also allows the config lock to be dropped. A
NULL config param, will trigger the lock take/drop within the routine, or
it will leave it untouched if passed in.

config_mount_ref, actually does the change, taking a lock. The second param
is really a boolean, does it increase or not.