change the source reserving api to not block
The reserving of the source structures require the source tree and source
lock with the source locked on exit. The return code now refers to any
blocking by either locks. It is unlikely that the source lock would block
unless it is a hijack situation.

source_client_startup becomes a restartable client handle after the client
was activated. This works in a similar way to the relay_initialise.