utilize the config lock changes on previoush commit. parent call re-ordering
Now that you can get a refcount on the mountinfo, to prevent the data
disappearing under you. This will allow for us to drop the config lock
while maintaining access to the mount structure if required.

Much of this really just looks like config lock, mount lookup and count,
config drop and then looking at the mount until you drop the reference.

The caller has to option to get the lock and pass it in, allowing for any
non-mount info to be retrieved before the lock is dropped.

The main one here is the source_recheck_mounts which was typically taking
the lock and traversing the tree, checking the source tree for active
sources. Now the config lock is only used to get to the next mount
definition. This is for cases with many mount definitions.  This does
mean that any assumption of the lock in the called functions is gone so
needed fixing up.

An awkward one was open_fh as it wants the directory paths to expand the
location, it also wants the fh_cache and these were annoyances for
lock tracking.

routines like fallback_count/fserve_query_count are non-blocking now, so
also had code path concerns.