Add priority setting onto relay host
introduce a priority flag onto hosts so that an order outside of the xml
can be indicated. The prority default of 0 implies an increase of the
priority value from the last on the list. The list added to in numerically
sorted order and same prorities are added afterwards. A value of 0 is
taken as meaning whatever the last one is +1.  This keeps the existing
setups the same.

The priorty value is not currently used so the only effect at the moment
is to allow hosts to be defined in the xml differently to as they were
before.