next update for switchover code.
This is to remove redundant parts and fix up some minor handling

If a source client is taking over a relay, then keep the existing relay,
just reset it and drop the client running it. No need to remove the
relay like we did with the host switch as that required a copy to
independently operate outside an already running relay.

Don't actually trigger a removal of the relay client in the switchover
case until the source client has been changed by new client code.
This allows for existing listeners to keep processing without being
concerned about the client disappearing