relay setup update
This is within the relay for reading the headers. Ever since the beginning
it has always assumed a non-encrypted stream. As such the routines used
were the original ones. With the recent http header work and ssl IO calls,
the update could now be done.

We use the same headers as before but you can specify a boolean option on
the host setting, either tls or ssl to enable encryption for that
connection. A redirect from a relay using https will also set it up.