fix relay issue, misalignment of icy blocks based on first read
Previously when reading relay header, it was taken 1 char at a time. Now
it is block based, we have to be careful on situation where stream data is
read with the headers. If it is then we need to filter those out the part
with the headers and stage it to be read later.  That way the correct
alignment is achieved.