auth fixes to avoid minor memory leak in certain cases
first is to use the aux_data to store the useagent for the mount_add/remove
triggers instead of the shared_data. There is no parser in those cases so
you cannot use that, and it was used pre-aux_data so use that as it is a
better fit for now. Could become a struct if more needed later. url auth
was the only auth to utilise this.

We used aux_data for holding the name for the command but would leak in
certain cases. The command is copied because it could of been aliased and
a delayed auth needs to reference it later in the auth thread.  In failure
cases the default send routines no nothing about the content of aux_data so
cannot touch it so would leak.  This instead pushes a reserved taken into
parser __admin_cnd and uses that for storage which is cleaned up
automatically during close.