clean up buffer block handling in relation to the http code
The http work allocates the block at the complete stage and inserts at the
head of the list for the client. This means that the original requirments
for block processing at the start has changed.  Before we had a block ready
to go when entering the request routines but now that is now dropped.

One to watch though is on incoming streams from sources. They can send
stream data after the connetion so you need to keep that around to avoid
any misalignment problems.