switchover code update
Next update on the routines. The mechanisms for switchover follow a pattern
of set the SOURCE_SWITCHOVER and set aux_data to refer to the new client
then wait for the original client to swap it over. The odd routine here is
the relay_reactivated routine which if it finds a relay for the source that
is terminating then there will be no relay client. In this case, we set
can set source client as there is no race between us, the listeners and the
another client.

The switchover routines for the relay client alway proceed to client free
up but it is just a matter of whether the relay is freed (priority host
switchover) or not (eg source override of relay).

With the linger-for setting, we can now have a source linger for a time
after termination and subsequent connection by a source client that
subsequent connection is treated as a switchover status and jumps in
meaning another mechanism for changing sources without a fallback

The previous update for switchover did not handle inactive on-demand
relays so now a source can override an on-demand relay and drop back
when finished

The linger-for setting does apply even in a multi-host relay. This is
something that may need addressing later.