fix speed assessment for fallback to file in certain cases
If no rate figure is set at source exit, then we need to make an assessment but
we need to do it in the shutdown not fallback as the fallback routine is only
called at the last stage which is not an issue for source clients but relays try
to restart immediately on exit so the connection time is reset.  By setting the
speed on the initial exit call, then any subsequent fails will have the rate
already primed.