slave thread operational markers simplified
The slave thread trigger various operations async from the workers and are set
via routines behind a spin lock. This copies them and then initiates those
opaerations outside of that spin lock.

This was convoluted before and some legacy code was remaining, making it look
like more was going on.