shoutcast compat case fix.
Originally the compat flag was explicitly used for sc clients. Then the
shoutcast-mount was placed in one block and that implicitly created the second
block. It was never really defined what happened if both were present, so I'll
assume that one with compat has it defined in the original way the the mount
will not do the second block. We also set the global if not already set.