config parse update, can set a default tag assignment
to allow for multiple ways for defining things, to maintain backward compliance
but allow for more settings which did not fit in the original layout.
eg errorlog only defined the name originally, with other tags that could affect
it like loglevel. later error log could be defined as a sub block allowing the
options but to allow both and the use of parameters, we can now nominate a tag
to be a default setting which would be the name for the error log, and other
settings can be specified as a block or parameters.