xml parsing update
this updates the handling of simple and nested tags for backward compatability.
First thing is we expand on the flag types for the tags. Introduce a default
field setting which allows for treating xml node data specially.  An example of
this is errorlog. The original setting just had a filename, but to allow for
other settings like level or size, it was made a group of tags but some tricks
were done to allow for either form to be accepted.  This allows us to use the
name element to be used in the original form.

Another tag introduced is to allow for multiple tags to be matched. Nothing is
using this currently to may prove useful later